donate help v4 (#247)
* Update social.py

Updated Donate helpfile
'Costs 5 AP' changed to 'Costs 1 AP'

* Automate Helpfile

Alters the donate helpfile so that it will automatically update the AP value the helpfile lists in the future, should the cost to run the command ever change again.

A slightly more complex test run to be sure I'm doing this right.

* Update social.py

* Revert "Update social.py"

This reverts commit 1f29713b60fa3c4755dd2b938fc2185d05ac6b53.

* Revert "Automate Helpfile"

This reverts commit a10a5f41440812781f31300a0f9cea4c828fcf30.

* Update social.py

* Update social.py

* Update home.py

* Update social.py

* Update home.py

* Update home.py

Reverted the change due to the Travis CI error. Not a valid way to write this solution.